Moved utf8 utils and tests
