+++ tests/index.js
@@ -29,7 +29,6 @@ import './unit/crypt/aes-decryptor';
 import './unit/crypt/decrypter';
 import './unit/demuxer/adts';
 import './unit/demuxer/base-audio-demuxer';
-import './unit/demuxer/id3';
 import './unit/demuxer/transmuxer';
 import './unit/loader/date-range';
 import './unit/loader/fragment-loader';
@@ -45,3 +44,4 @@ import './unit/utils/exp-golomb';
 import './unit/utils/output-filter';
 import './unit/utils/texttrack-utils';
 import './unit/utils/vttparser';
+import './unit/utils/utf8';
